refactor drag and drop context
