+++ src/App.tsx
@@ -6,22 +6,25 @@ import AboutPage from "./pages/About";
 import { OpenSidebarProvider } from "./context/SidebarContext";
 import { Toaster } from "./components/@/components/ui/toaster";
 import { KanbanProvider } from "./context/KanbanContext";
+import { DragAndDropProvider } from "./context/DragAndDropContext";
 
 function App() {
   return (
     <>
       <KanbanProvider>
-        <OpenSidebarProvider>
-          <BrowserRouter>
-            <Routes>
-              <Route element={<AppLayout />}>
-                <Route path="/" element={<Homepage />} />
-                <Route path="/tasks" element={<TaskPage />} />
-                <Route path="/about" element={<AboutPage />} />
-              </Route>
-            </Routes>
-          </BrowserRouter>
-        </OpenSidebarProvider>
+        <DragAndDropProvider>
+          <OpenSidebarProvider>
+            <BrowserRouter>
+              <Routes>
+                <Route element={<AppLayout />}>
+                  <Route path="/" element={<Homepage />} />
+                  <Route path="/tasks" element={<TaskPage />} />
+                  <Route path="/about" element={<AboutPage />} />
+                </Route>
+              </Routes>
+            </BrowserRouter>
+          </OpenSidebarProvider>
+        </DragAndDropProvider>
       </KanbanProvider>
       <Toaster />
     </>
